add partial implementation for generic_event event logger
