+++ rlclientlib/logger/logger_extensions.h
@@ -1,0 +1,44 @@
+#pragma once
+
+#include "configuration.h"
+#include "error_callback_fn.h"
+
+#include <functional>
+#include <string>
+
+namespace reinforcement_learning{
+// forward declare all the types
+namespace utility { class watchdog; }
+class generic_event;
+class api_status;
+class i_time_provider;
+namespace logger { class i_message_sender; }
+
+namespace logger {
+
+    // WARNING: This interface is a bit complex in its usage. It currently lives in the live_model_impl, but is passed
+    // into the interaction logger and is eventually called in the logger thread. However, it (currently) does not
+    // need any form of locking because the object is ONLY called in the SINGLE logger thead.
+    // The workflow looks like:
+    //   live_model_impl (owner) CONTAINS interaction_logger_facade CONTAINS generic_event_logger CONTAINS
+    //     async_batcher CONTROLS logger thread AND CONTAINS a queue of generic_event. generic_event will hold a pointer to this object
+    // If this condition changes, then the relevant calls must be locked!
+    class i_logger_extensions {
+    protected:
+      const utility::configuration& _config;
+    public:
+      i_logger_extensions(const utility::configuration&);
+
+      virtual ~i_logger_extensions();
+
+      virtual bool is_object_extraction_enabled() const = 0;
+      virtual bool is_serialization_transform_enabled() const = 0;
+
+      virtual i_async_batcher<std::function<int(generic_event&, api_status*)>>* create_batcher(i_message_sender* sender, utility::watchdog& watchdog, error_callback_fn* perror_cb, const char* section) = 0;
+      virtual int transform_payload_and_extract_objects(const char* context, std::string& edited_payload, generic_event::object_list_t& objects, api_status* status) = 0;
+      virtual int transform_serialized_payload(generic_event::payload_buffer_t& input, event_content_type &content_type, api_status* status) const = 0;
+
+      static i_logger_extensions* get_extensions(const utility::configuration& config, i_time_provider* time_provider);
+    };
+
+}}
